LayoutTests/fast/events/keydown-numpad-keys-expected.txt


This tests keyboard events with KeyLocationCode argument.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 33
PASS lastKeyboardEvent.location is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 34
PASS lastKeyboardEvent.location is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 36
PASS lastKeyboardEvent.location is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 35
PASS lastKeyboardEvent.location is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 37
PASS lastKeyboardEvent.location is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 39
PASS lastKeyboardEvent.location is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 38
PASS lastKeyboardEvent.location is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 40
PASS lastKeyboardEvent.location is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 45
PASS lastKeyboardEvent.location is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 46
PASS lastKeyboardEvent.location is 0
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 33
PASS lastKeyboardEvent.location is 3
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 34
PASS lastKeyboardEvent.location is 3
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 36
PASS lastKeyboardEvent.location is 3
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 35
PASS lastKeyboardEvent.location is 3
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 37
PASS lastKeyboardEvent.location is 3
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 39
PASS lastKeyboardEvent.location is 3
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 38
PASS lastKeyboardEvent.location is 3
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 40
PASS lastKeyboardEvent.location is 3
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 45
PASS lastKeyboardEvent.location is 3
PASS lastKeyboardEvent.type is "keydown"
PASS lastKeyboardEvent.keyCode is 46
PASS lastKeyboardEvent.location is 3
PASS successfullyParsed is true

TEST COMPLETE
